When starting to parse a number assume integer
In State.ValueBegin and State.ValueBeginNoClosing when the character is
'-', '0' or '1'...'9' p.number_is_integer is set to true. This allows
new numeric values to be either Value.Integer or Value.Float. Otherwise
when a Float is encountered all subsequent numbers will be Value.Float's.

This was found while enhancing the json parser[1] to support floats and
adding tests like the following. Without this change the test fails
because "b" is a Value.Float rather than the expected Value.Integer:

 test "json.simple.numbers" {
    var p = Parser.init(debug.global_allocator, false);
    defer p.deinit();

    const s =
        \\{"a": -1.0, "b": 2}
    ;

    var tree = try p.parse(s);
    defer tree.deinit();

    var root = tree.root;

    debug.assert(root.Object.get("a").?.value.Float == -1.0);
    debug.assert(root.Object.get("b").?.value.Integer == 2);
 }


[1]: https://github.com/winksaville/zig-json